refactor: 위치 출력하는 테스트 PositionTest 클래스에서 수행 및 중복되는 테스트 삭제
