+++ src/test/java/racingcar/CarTest.java
@@ -1,6 +1,5 @@
 package racingcar;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -23,27 +22,11 @@ class CarTest {
     }
 
     @ParameterizedTest
-    @ValueSource(ints = {3})
-    @DisplayName("자동차의 출력 될 상태를 확인한다.")
-    void checkPosition(int position) {
-        Car car = new Car("jhLim",position);
-
-        assertThat(car.getPosition()).isEqualTo(new Position(position));
-    }
-
-    @ParameterizedTest
     @ValueSource(ints = {-1, 10})
     @DisplayName("0에서 9사이의 숫자 이외에는 자동차가 판단할 수 없다.")
     void checkRange(int number) {
         assertThatIllegalArgumentException().isThrownBy(() -> {
             new Car("jhLim97").conditionValueValidation(number);
         });
-    }
-
-    @ParameterizedTest
-    @CsvSource({"3, ---"})
-    @DisplayName("자동차의 위치를 표시할 수 있다.")
-    void canDisplayCarPosition(int position, String displayPosition) {
-        Assertions.assertThat(new Car("jhLim", position).toString()).isEqualTo(displayPosition);
     }
 }
